♻️ refactor: extract shared OptionValue/normalizeOptions for Radio/Checkbox groups
RadioGroup and CheckboxGroup each redefined the same OptionValue/
Option/Options types and the same options-normalization logic.
Pulled just that shared, truly-identical part into
atoms/option-group.ts — selection semantics (single-select Radix
root vs multi-select array) stay separate since they're not actually
the same logic.

Also tightens both groups' `classNames` from a loose
`Record<string, string>` to the `{ wrapper?; item? }` shape they
already only ever use.
